Add some test coverage to callsets for searching by name. fix #142
