+++ cts-java/src/test/java/org/ga4gh/cts/api/TestData.java
@@ -129,6 +129,12 @@ public class TestData {
      */
     public static final String ALIGNED_SEQUENCE_CONTENTS_PATTERN = "[ACTGN]+";
 
+
+    /**
+     * The call set name is used to test searching by call set name functionality.
+     */
+    public static final String CALL_SET_NAME = "HG00096";
+
     /**
      * Return the ID of the compliance dataset on the server being tested.
      * By default this is the value of {@link #DEFAULT_DATASET_ID}, but
